requirements.txt
psycopg2
tkcalendar
Pillow
reportlab
